affine (#29)
* generalize AffineFor (WIP)

* 'getTypeAtIndex' wants unsigned as argument

* Handle affine fors generally

* Fix a few Wreorder warnings (NFC)

* Clang-format (NFC)

* add canonicalization patter to detect cast from index to index

* Drop broken cast

The cast was pushing an 'index' type in the affine loop descriptor type, which
is wrong as an affine loop descriptor type should follow the clang type.

* Clang-format

* move Canonicalization pass after AffineLoopInvariantCodeMotion

* Revert "Emit affine.store for add and mul assignment operators"

This reverts commit 745fa83608cd9b7ef6e3d5820ebe708c1395c064. A next commit
will re-enable emission of affine.store but following a different approach.

* Handle LE and fix parallel bug

* fix emission for LE

* Clean up and start negation

* Fix affine rewriting

* start reworking tests (WIP)

* clang format

* Make test fail as we don't check the IR for now

* Emit affine.store and affine.load

Introduce a function pass to replace store with affine.store/affine.load if the
arguments are indexes of an affine.for or are constants.

Co-authored-by: William S. Moses <gh@wsmoses.com>